requirements.txt
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
joblib==1.4.2
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.5.0
scipy==1.13.1
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.0
svgwrite==1.4.3
threadpoolctl==3.5.0
Tree==0.2.4
tzdata==2024.1
